requirements.txt
python-consul==1.1.0
hvac==2.1.0
pyyaml==6.0.1
